Based on the review feedback. the API for uploading files using the knowledge base name has been modified.
